create_AgProdChange_xml

Function that takes the defined impacts and reference AgProdChange files then updates and saves the new AgProdChange csv and xml files.

create_AgProdChange_xml(
  write_dir = "outputs_create_AgProdChange_xml",
  ag_irr_ref = NULL,
  bio_irr_ref = NULL,
  ag_impacts = NULL,
  bio_impacts = NULL,
  GCAM_region_mapping = NULL,
  timestep = 5,
  maxHistYear = 2010,
  minFutYear = 2015,
  appliedto = "full"
)

Arguments

write_dir

Default = "outputs_create_AgProdChange_xml". Output Folder

ag_irr_ref

Default = NULL

bio_irr_ref

Default = NULL

ag_impacts

Default = NULL

bio_impacts

Default = NULL

GCAM_region_mapping

Default = NULL

timestep

Default = 5

maxHistYear

Default = 2010 Historical year

minFutYear

Default = 2015 Min future year

appliedto

Default = "full". This applies impacts to all regions. Other options are "domestic" and "international".

Value

number

Examples

if (FALSE) {
library(osiris)
osiris::create_AgProdChange_xml()
}