grid_to_basin_yield

Function that uses emulators to create gridded yield files then aggregates to gcam basin level via MIRCA harvested areas.

grid_to_basin_yield(
  carbon = NULL,
  weight_floor_ha = 1,
  emulator_dir = NULL,
  input_dir = NULL,
  area_dir = NULL,
  basin_grid = NULL,
  basin_id = NULL,
  region_id = NULL,
  gridded_yield_dir = NULL,
  write_dir = "step2_grid_to_basin_yield",
  crops = c("Corn", "Rice", "Soy", "Spring Wheat", "Winter Wheat"),
  esm_name = "WRF",
  cm_name = "LPJmL",
  scn_name = "rcp8p5_hotter",
  N = 200
)

Arguments

carbon

Default = NULL

weight_floor_ha

Default = 1

emulator_dir

Default = NULL

input_dir

Default = NULL

area_dir

Default = NULL

basin_grid

Default = NULL

basin_id

Default = NULL

region_id

Default = NULL. Filters to specified GCAM region (1-32), otherwise no filter

gridded_yield_dir

Default = NULL.

write_dir

Default = "step2_grid_to_basin_yield". Output Folder

crops

Default = c("Corn", "Rice", "Soy", "Spring Wheat", "Winter Wheat")

esm_name

Default = 'WRF'

cm_name

Default = 'LPJmL'

scn_name

Default = 'rcp8p5_hotter'

N

Default = 200. Assuming nothing is nitrogen limited and apply across grids

Value

number

Examples

if (FALSE) {
library(osiris)
osiris::grid_to_basin_yield()
}