yield_to_gcam_basin

Function that uses emulated yield data aggregated to GCAM basin level to create GCAM region, basin, and irrigation level data for all GCAM commodities

yield_to_gcam_basin(
  write_dir = "outputs_yield_to_gcam_basin",
  emulated_basin_yield_dir = NULL,
  iso_GCAM_region_mapping = NULL,
  FAO_ag_mapping = NULL,
  iso_harvest_area_mapping = NULL,
  iso_GCAM_basin_mapping = NULL,
  esm_name = "WRF",
  scn_name = "rcp8p5_hot",
  max_CCImult = 2.5,
  min_CCImult = 0.01,
  weight_floor_ha = 1,
  rolling_avg_years = 15,
  maxHistYear = 2010,
  minFutYear = 2015,
  maxFutYear = 2100,
  extrapolate_to = NULL
)

Arguments

write_dir

Default = "outputs_yield_to_gcam_basin". Output Folder

emulated_basin_yield_dir

Default = NULL

iso_GCAM_region_mapping

Default = NULL

FAO_ag_mapping

Default = NULL

iso_harvest_area_mapping

Default = NULL

iso_GCAM_basin_mapping

Default = NULL

esm_name

Default = 'WRF'

scn_name

Default = 'rcp8p5_hot'

max_CCImult

Default = 2.5 Upper limit on positive climate impacts (multiplier)

min_CCImult

Default = 0.01 Lower limit on negative climate impacts (multiplier)

weight_floor_ha

Default = 1 Floor on area weights, in hectares. Below this climate impacts will be ignored. These are more likely than others to be problematic. 1 hectare = 0.01 km^2 = 1e-5 thou km^2, GCAM land units.

rolling_avg_years

Default = 15 Set the number of years to define the range for rolling averages (range = 2 times this plus 1)

maxHistYear

Default = 2010 Historical year for which to apply rolling averages

minFutYear

Default = 2015 Min future year for which to apply rolling averages

maxFutYear

Default = 2100 Max future year for which to apply rolling averages

extrapolate_to

Default = NULL. Note that this extrapolates to only one future year (eg, 2099 to 2100)

Value

number

Examples

if (FALSE) {
library(osiris)
osiris::yield_to_gcam_basin()
}